# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Reservations API",
        "description": "API REST para gestão de hotéis, quartos e reservas (desafio técnico Foco Multimídia). A consulta de quartos é pública. Cadastro, edição e exclusão de quartos e todas as rotas de reservas exigem token: faça login em POST /login e use o botão Authorize.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor local (Docker)"
        }
    ],
    "paths": {
...
        "/rooms": {
            "get": {
                "tags": [
                    "Quartos"
                ],
                "summary": "Lista os quartos",
                "operationId": "1b492e14cbdc859a446992c209170929",
                "parameters": [
                    {
                        "name": "hotel_id",
                        "in": "query",
                        "description": "Filtra os quartos de um hotel",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página (15 itens por página)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de quartos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Room"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Quartos"
                ],
                "summary": "Cadastra um quarto",
                "operationId": "0fcc821623ab708a9cd94310cad7971f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "hotel_id",
                                    "name"
                                ],
                                "properties": {
                                    "hotel_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Suite Master"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Quarto criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Room"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados inválidos (hotel inexistente, nome ausente ou repetido no mesmo hotel)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/rooms/{id}": {
            "get": {
                "tags": [
                    "Quartos"
                ],
                "summary": "Detalha um quarto",
                "operationId": "6648acdecabaa361b4a002000e0f9c47",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do quarto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Quarto encontrado, com o hotel",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Room"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Quarto não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Quartos"
                ],
                "summary": "Atualiza o nome de um quarto",
                "description": "Apenas o nome pode ser alterado. O nome deve ser único dentro do hotel.",
                "operationId": "373c3c7b58c51231b14ea7af07ee9036",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do quarto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Suite Luxo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Quarto atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Room"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Quarto não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Nome ausente ou já usado por outro quarto do mesmo hotel",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Quartos"
                ],
                "summary": "Exclui um quarto",
                "description": "Quartos que possuem reservas não podem ser excluídos.",
                "operationId": "cc27db8b69e5e1d2df63959e66ac4ada",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do quarto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Quarto excluído (sem conteúdo)"
                    },
                    "404": {
                        "description": "Quarto não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "O quarto possui reservas e não pode ser excluído",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Room": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "hotel_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Room 1 Hotel 1"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Cole apenas o token devolvido pelo login (sem a palavra Bearer).",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}